improve files organization
