+++ lib/src/pluto_grid.dart
@@ -1240,220 +1240,6 @@ class _GridContainer extends StatelessWidget {
   }
 }
 
-/// [PlutoGrid.onLoaded] Argument received by registering callback.
-class PlutoGridOnLoadedEvent {
-  final PlutoGridStateManager stateManager;
-
-  const PlutoGridOnLoadedEvent({
-    required this.stateManager,
-  });
-}
-
-/// Event called when the value of [PlutoCell] is changed.
-///
-/// Notice.
-/// [columnIdx], [rowIdx] are the values in the current screen state.
-/// Values in their current state, not actual data values
-/// with filtering, sorting, or pagination applied.
-/// This value is from
-/// [PlutoGridStateManager.columns] and [PlutoGridStateManager.rows].
-///
-/// All data is in
-/// [PlutoGridStateManager.refColumns.originalList]
-/// [PlutoGridStateManager.refRows.originalList]
-class PlutoGridOnChangedEvent {
-  final int columnIdx;
-  final PlutoColumn column;
-  final int rowIdx;
-  final PlutoRow row;
-  final dynamic value;
-  final dynamic oldValue;
-
-  const PlutoGridOnChangedEvent({
-    required this.columnIdx,
-    required this.column,
-    required this.rowIdx,
-    required this.row,
-    this.value,
-    this.oldValue,
-  });
-
-  @override
-  String toString() {
-    String out = '[PlutoOnChangedEvent] ';
-    out += 'ColumnIndex : $columnIdx, RowIndex : $rowIdx\n';
-    out += '::: oldValue : $oldValue\n';
-    out += '::: newValue : $value';
-    return out;
-  }
-}
-
-/// This is the argument value of the [PlutoGrid.onSelected] callback
-/// that is called when the [PlutoGrid.mode] value is in select mode.
-///
-/// If [row], [rowIdx], [cell] is [PlutoGridMode.select] or [PlutoGridMode.selectWithOneTap],
-/// Information of the row selected with the tab or enter key.
-/// If the Escape key is pressed, these values are null.
-///
-/// [selectedRows] is valid only in case of [PlutoGridMode.multiSelect].
-/// If rows are selected by tab or keyboard, the selected rows are included.
-/// If the Escape key is pressed, this value is null.
-class PlutoGridOnSelectedEvent {
-  final PlutoRow? row;
-  final int? rowIdx;
-  final PlutoCell? cell;
-  final List<PlutoRow>? selectedRows;
-
-  const PlutoGridOnSelectedEvent({
-    this.row,
-    this.rowIdx,
-    this.cell,
-    this.selectedRows,
-  });
-
-  @override
-  String toString() {
-    return '[PlutoGridOnSelectedEvent] rowIdx: $rowIdx, selectedRows: ${selectedRows?.length}';
-  }
-}
-
-/// Argument of [PlutoGrid.onSorted] callback for receiving column sort change event.
-class PlutoGridOnSortedEvent {
-  final PlutoColumn column;
-
-  final PlutoColumnSort oldSort;
-
-  const PlutoGridOnSortedEvent({
-    required this.column,
-    required this.oldSort,
-  });
-
-  @override
-  String toString() {
-    return '[PlutoGridOnSortedEvent] ${column.title} (changed: ${column.sort}, old: $oldSort)';
-  }
-}
-
-/// Argument of [PlutoGrid.onRowChecked] callback to receive row checkbox event.
-///
-/// [runtimeType] is [PlutoGridOnRowCheckedAllEvent] if [isAll] is true.
-/// When [isAll] is true, it means the entire check button event of the column.
-///
-/// [runtimeType] is [PlutoGridOnRowCheckedOneEvent] if [isRow] is true.
-/// If [isRow] is true, it means the check button event of a specific row.
-abstract class PlutoGridOnRowCheckedEvent {
-  bool get isAll => runtimeType == PlutoGridOnRowCheckedAllEvent;
-
-  bool get isRow => runtimeType == PlutoGridOnRowCheckedOneEvent;
-
-  final PlutoRow? row;
-  final int? rowIdx;
-  final bool? isChecked;
-
-  const PlutoGridOnRowCheckedEvent({
-    this.row,
-    this.rowIdx,
-    this.isChecked,
-  });
-
-  @override
-  String toString() {
-    String checkMessage = isAll ? 'All rows ' : 'RowIdx $rowIdx ';
-    checkMessage += isChecked == true ? 'checked' : 'unchecked';
-    return '[PlutoGridOnRowCheckedEvent] $checkMessage';
-  }
-}
-
-/// Argument of [PlutoGrid.onRowChecked] callback when the checkbox of the row is tapped.
-class PlutoGridOnRowCheckedOneEvent extends PlutoGridOnRowCheckedEvent {
-  const PlutoGridOnRowCheckedOneEvent({
-    required PlutoRow super.row,
-    required int super.rowIdx,
-    required super.isChecked,
-  });
-}
-
-/// Argument of [PlutoGrid.onRowChecked] callback when all checkboxes of the column are tapped.
-class PlutoGridOnRowCheckedAllEvent extends PlutoGridOnRowCheckedEvent {
-  const PlutoGridOnRowCheckedAllEvent({
-    super.isChecked,
-  }) : super(row: null, rowIdx: null);
-}
-
-/// The argument of the [PlutoGrid.onRowDoubleTap] callback
-/// to receive the event of double-tapping the row.
-class PlutoGridOnRowDoubleTapEvent {
-  final PlutoRow row;
-  final int rowIdx;
-  final PlutoCell cell;
-
-  const PlutoGridOnRowDoubleTapEvent({
-    required this.row,
-    required this.rowIdx,
-    required this.cell,
-  });
-}
-
-/// Argument of the [PlutoGrid.onRowSecondaryTap] callback
-/// to receive the event of tapping the row with the right mouse button.
-class PlutoGridOnRowSecondaryTapEvent {
-  final PlutoRow row;
-  final int rowIdx;
-  final PlutoCell cell;
-  final Offset offset;
-
-  const PlutoGridOnRowSecondaryTapEvent({
-    required this.row,
-    required this.rowIdx,
-    required this.cell,
-    required this.offset,
-  });
-}
-
-/// Argument of [PlutoGrid.onRowsMoved] callback
-/// to receive the event of moving the row by dragging it.
-class PlutoGridOnRowsMovedEvent {
-  final int idx;
-  final List<PlutoRow> rows;
-
-  const PlutoGridOnRowsMovedEvent({
-    required this.idx,
-    required this.rows,
-  });
-}
-
-/// Argument of [PlutoGrid.onColumnsMoved] callback
-/// to move columns by dragging or receive left or right fixed events.
-///
-/// [idx] means the actual index of
-/// [PlutoGridStateManager.columns] or [PlutoGridStateManager.refColumns].
-///
-/// [visualIdx] means the order displayed on the screen, not the actual index.
-/// For example, if there are 5 columns of [0, 1, 2, 3, 4]
-/// If 1 column is frozen to the right, [visualIndex] becomes 4.
-/// But the actual index is preserved.
-class PlutoGridOnColumnsMovedEvent {
-  final int idx;
-  final int visualIdx;
-  final List<PlutoColumn> columns;
-
-  const PlutoGridOnColumnsMovedEvent({
-    required this.idx,
-    required this.visualIdx,
-    required this.columns,
-  });
-
-  @override
-  String toString() {
-    String text =
-        '[PlutoGridOnColumnsMovedEvent] idx: $idx, visualIdx: $visualIdx\n';
-
-    text += columns.map((e) => e.title).join(',');
-
-    return text;
-  }
-}
-
 /// Argument of [PlutoGrid.rowColumnCallback] callback
 /// to dynamically change the background color of a row.
 class PlutoRowColorContext {
@@ -1514,187 +1300,6 @@ class PlutoOptional<T> {
   const PlutoOptional(this.value);
 
   final T? value;
-}
-
-abstract class PlutoGridSettings {
-  /// If there is a frozen column, the minimum width of the body
-  /// (if it is less than the value, the frozen column is released)
-  static const double bodyMinWidth = 200.0;
-
-  /// Default column width
-  static const double columnWidth = 200.0;
-
-  /// Column width
-  static const double minColumnWidth = 80.0;
-
-  /// Frozen column division line (ShadowLine) size
-  static const double shadowLineSize = 3.0;
-
-  /// Sum of frozen column division line width
-  static const double totalShadowLineWidth =
-      PlutoGridSettings.shadowLineSize * 2;
-
-  /// Grid - padding
-  static const double gridPadding = 2.0;
-
-  /// Grid - border width
-  static const double gridBorderWidth = 1.0;
-
-  static const double gridInnerSpacing =
-      (gridPadding * 2) + (gridBorderWidth * 2);
-
-  /// Row - Default row height
-  static const double rowHeight = 45.0;
-
-  /// Row - border width
-  static const double rowBorderWidth = 1.0;
-
-  /// Row - total height
-  static const double rowTotalHeight = rowHeight + rowBorderWidth;
-
-  /// Cell - padding
-  static const EdgeInsets cellPadding = EdgeInsets.symmetric(horizontal: 10);
-
-  /// Column title - padding
-  static const EdgeInsets columnTitlePadding =
-      EdgeInsets.symmetric(horizontal: 10);
-
-  static const EdgeInsets columnFilterPadding = EdgeInsets.all(5);
-
-  /// Cell - fontSize
-  static const double cellFontSize = 14;
-
-  /// Scroll when multi-selection is as close as that value from the edge
-  static const double offsetScrollingFromEdge = 10.0;
-
-  /// Size that scrolls from the edge at once when selecting multiple
-  static const double offsetScrollingFromEdgeAtOnce = 200.0;
-
-  static const int debounceMillisecondsForColumnFilter = 300;
-}
-
-enum PlutoGridMode {
-  /// {@template pluto_grid_mode_normal}
-  /// Basic mode with most functions not limited, such as editing and selection.
-  /// {@endtemplate}
-  normal,
-
-  /// {@template pluto_grid_mode_readOnly}
-  /// Cell cannot be edited.
-  /// To try to edit by force, it is possible as follows.
-  ///
-  /// ```dart
-  /// stateManager.changeCellValue(
-  ///   stateManager.currentCell!,
-  ///   'test',
-  ///   force: true,
-  /// );
-  /// ```
-  /// {@endtemplate}
-  readOnly,
-
-  /// {@template pluto_grid_mode_select}
-  /// Mode for selecting one list from a specific list.
-  /// Tap a row or press Enter to select the current row.
-  ///
-  /// [select]
-  /// Call the [PlutoGrid.onSelected] callback when the selected row is tapped.
-  /// To select an unselected row, select the row and then tap once more.
-  /// [selectWithOneTap]
-  /// Same as [select], but calls [PlutoGrid.onSelected] with one tap.
-  ///
-  /// This mode is non-editable, but programmatically possible.
-  /// ```dart
-  /// stateManager.changeCellValue(
-  ///   stateManager.currentRow!.cells['column_1']!,
-  ///   value,
-  ///   force: true,
-  /// );
-  /// ```
-  /// {@endtemplate}
-  select,
-
-  /// {@macro pluto_grid_mode_select}
-  selectWithOneTap,
-
-  /// {@template pluto_grid_mode_multiSelect}
-  /// Mode to select multiple rows.
-  /// When a row is tapped, it is selected or deselected and the [PlutoGrid.onSelected] callback is called.
-  /// [PlutoGridOnSelectedEvent.selectedRows] contains the selected rows.
-  /// When a row is selected with keyboard shift + arrowDown/Up keys,
-  /// the [PlutoGrid.onSelected] callback is called only when the Enter key is pressed.
-  /// When the Escape key is pressed,
-  /// the selected row is canceled and the [PlutoGrid.onSelected] callback is called
-  /// with a [PlutoGridOnSelectedEvent.selectedRows] value of null.
-  /// {@endtemplate}
-  multiSelect,
-
-  /// {@template pluto_grid_mode_popup}
-  /// This is a mode for popup type.
-  /// It is used when calling a popup for filtering or column setting
-  /// inside [PlutoGrid], and it is not a mode for users.
-  ///
-  /// If the user wants to run [PlutoGrid] as a popup,
-  /// use [PlutoGridPopup] or [PlutoGridDualGridPopup].
-  /// {@endtemplate}
-  popup;
-
-  bool get isNormal => this == PlutoGridMode.normal;
-
-  bool get isReadOnly => this == PlutoGridMode.readOnly;
-
-  bool get isEditableMode => isNormal || isPopup;
-
-  bool get isSelectMode => isSingleSelectMode || isMultiSelectMode;
-
-  bool get isSingleSelectMode => isSelect || isSelectWithOneTap;
-
-  bool get isMultiSelectMode => isMultiSelect;
-
-  bool get isSelect => this == PlutoGridMode.select;
-
-  bool get isSelectWithOneTap => this == PlutoGridMode.selectWithOneTap;
-
-  bool get isMultiSelect => this == PlutoGridMode.multiSelect;
-
-  bool get isPopup => this == PlutoGridMode.popup;
-}
-
-/// When calling loading screen with [PlutoGridStateManager.setShowLoading] method
-/// Determines the level of loading.
-///
-/// {@template pluto_grid_loading_level_grid}
-/// [grid] makes the entire grid opaque and puts the loading indicator in the center.
-/// The user is in a state where no interaction is possible.
-/// {@endtemplate}
-///
-/// {@template pluto_grid_loading_level_rows}
-/// [rows] represents the [LinearProgressIndicator] at the top of the widget area
-/// that displays the rows.
-/// User can interact.
-/// {@endtemplate}
-///
-/// {@template pluto_grid_loading_level_rowsBottomCircular}
-/// [rowsBottomCircular] represents the [CircularProgressIndicator] at the bottom of the widget
-/// that displays the rows.
-/// User can interact.
-/// {@endtemplate}
-enum PlutoGridLoadingLevel {
-  /// {@macro pluto_grid_loading_level_grid}
-  grid,
-
-  /// {@macro pluto_grid_loading_level_rows}
-  rows,
-
-  /// {@macro pluto_grid_loading_level_rowsBottomCircular}
-  rowsBottomCircular;
-
-  bool get isGrid => this == PlutoGridLoadingLevel.grid;
-
-  bool get isRows => this == PlutoGridLoadingLevel.rows;
-
-  bool get isRowsBottomCircular =>
-      this == PlutoGridLoadingLevel.rowsBottomCircular;
 }
 
 enum _StackName {
